Cortana (at least Redstone): silence name change event when listening mode is in progress.
A more reliable way found: when user is speaking to Cortana, search edit field disappears. This can be advantageous, as search field is right next to dictation text, so it is possible to suppress announcing gibberish.
As this requires testing in both 1511 and 1607, it'll not make it to stable until October.