UIA: Add event handler requests nondestructively.
Until now, UIAHandler had to be restarted in order to let it recognize new events to process. However, it was found that one can add event handlers directly without a need to restart UIA handler thread. Even then, it is necessary to add event name key and value to events map.
This solution also resolves timeout errors seen on some systems when the add-on loads.