Settings app: greatly reorganized name change event that surprisingly works with 1511 and 1607.
Instead of checking if there is a progress bar next door, use automation ID's to weed out what should not be announced. For now, Windows Update notification, search progress, and search error will be announced, with ID's based on build 14393 (and works in 10586). Note that UIA in version 1511 is quite bad (quits after using nVDA for a while).