Skype Preview: foundation to allow more commands and status to be announced.
For now, one of the reliable ways to obtain elements is scraping the screen (for loop). As of now, a dedicated function has been added to locate the desired element (based on automation ID), useful for locating chat history, edit fields, contacts list and others (in hopes of adding keyboard shortcuts).