Windows Defender UWP: Add a workaround to let NVDA announce button labels.
Build 14986 and later: similar to Twitter and other app modules, current version of Windows Defender UWP has no labels for certain buttons. Thankfully, the last child of this control has the labels we need, so fetch this instead.
This is a temporary solution and will be removed once correct labels from Microsoft are released and implemented correctly.