Windows Store: announce app and other content download progress.
While an app or content is downloaindg, a progress text is shown that provides useful info for users to hear, including download status, size and so on. However, because it also suffers from cache issue (same thing spoken multiple times), provide a cache so different info can be announced. Also, product title will be announced along with the progress text.