Settings app (15000 and later): provide a heuristic way of locating combo boxes where there are no labels.
Certain combo boxes in Settings app are not labeled (example: Apps/Apps and Features/app location in build 15000 and later). There is a handy way to locate labels for these: they have automation ID's and the label is right next to the combo box in the object hierarchy. Thus use a heuristic to locate the label.