Update check: refinements, data structures no longer defined, logging improvements.
After downloading and installing a new add-on version, make sure to delete the bundle from temp path.
While logging, do not refer to a nonexistent variable.
Certain structures that were part of the w10config module will not be present, as they'll be pulled in straight from NVDA Core's update check module.