Maps: let NVDA announce street side information when navigating via keyboard. re #18
Requested by a user: in Maps app, it is possible to look at street side and use the arrow keys to walk the streets, and NVDA did not announce the new location (now it does). This is thanks to the fact that the street side text block has its own UIA automation ID and fires name change event.