Controller for workaround (again): support at mentions in Mail app, as it also raises controller for event.
In Mail's message window, when typing the at sign (@), a list of contacts will appear. This is called 'at mentions' and once this happens, a controller for event is fired by the message composition window. Due to current controller for handler being optimized for search fields, a new UIAEditableTextWithSuggestions is hereby introduced to handle search fields and composition windows.