People app/My People: allow suggestions sound to be heard when contacts appear.
In People and My Peopple (Redstone 3), a search box is provided to search for contacts. Although this field does not raise controller for event, when contacts appear, it shows up as list items in UIA, so controller for event can be emulated (the catch here is checking to make sure that value change produced some obj.value data).