Edge: add support for announcing live text alerts.
Edge fires live region change event when alerts appear, consistent with some browsers such as Firefox. It's been difficult to locate which element that's doing this, but what's more odd is the fact that obj.name is empty yet the child exposes the actual message. Thus use this to announce alerts in Edge such as ad skipping notices on YouTube.