Calculator: prevent double focus announcements in some cases, screen layout change.
In newer releases, calculator result is shown on the same spot on the object hierarchy, thus take a note of this. This necessarily means newest Calculator app release is supported.
When handling escape and enter keys, just look at automation ID. This allows controls other than radio buttons to be recognized when deaing with these commands, which allows display text to be announced no matter where one is at.