Build 162xx/Action Center: a hack to allow tab key to work when notification list is empty.
In recent 162xx builds, when pressing tab from Action Center with no notifications, it does not permit movement to other parts such as quick actions. Partially restore this thorugh a scripted hack (this is a temporary workaround and hopefully to be gone by the time Fall Creators Update is released).