Settings app: yet another workaround for feature update entry in update history.
In build 16251 and later (or perhaps 16288), feature update object is a link. Thankfully, the previous.previous object and the previous ones have the same event ID, so one can construct the UIA automation ID and compare it against the actual ID for the feature update text (and it works, but can change in a future release).