Settings: announce more Windows Update status messages.
In build 17063, update status is split between main text bock and a subtitle. Due to complexity of live region checks, this is now part of a brand new function.
Added announceLiveRegion test function that now returns if the new live region should be announced. This also includes a bit complicated ternary operation (necessary due to conditions and the fact that progress bars should not be announced in Windows Update unless have to).