UIA event tracking: track UIA notification event introduced in Version 1709.
In Windows 10 Version 1709 (Fall Creators Update), a new notification event is available for use by apps to communicate with screen readers for announcing various things. So far, Calculator fires this event when entries change, and it could be expanded to other apps later.
Unfortuantely, this meant resurrecting manual event insertion seen in older verions of this add-on. At least, preparation has been made to track and respond to this event. For now, it'll be cataloged under UIA_notification event.