Virtual desktops: detect the workaround from NVDA Core.
NVDA Core issue 5641: a new event handler will be introduced that'll detect name change event for the desktop object. Note that this handler will be part of NVDA Core (not the add-on itself, although parts of it will be included to let older NVDA releases handle this in a consistent manner).