IUIA6: add isDialog property and detection debug tone.
In Windows 10 Redstone 5, IUIAutoamtionElement9 includes a new IsDialog property (30174) to let screen readers and others know that they are dealing with a dialog element. Because not all dialogs are detected as dialogs, allow NVDA to at least try detecting dialogs during overlay class chooser (and play a tone if found).