IUIAutomationElement9/IsDialog: use this property to check if windows are relaly dialogs, applicable in build 17682 and later.
Thanks to IsDialog property in IUIAutomationElement9 (build 17682 and later), it became easier to check if a window is really a dialog in disguise. However, just checking this isn't enough: people are using old Windows 10 releases such as Version 1709, so make sure to detect dialogs using UIA class name method. This code also serves as a test for NVDA Core issue 8405.