Update check: turn off update check module (w10config) if updating is not supported.
Whenever add-on updating isn't supported, raise runtime errors when trying to import w10config (that module's only job was to configure and check for add-on update feature anyway).
This commit also lays the foundation to let Add-on Updater (proof of concept) check for Windows 10 App Essentials updates in the future.