Calculator: announce unit conversion results again.
Thanks to UI redesign in July 2018, Calculator app module raises attribute error because the notification object is a singleton (no next, no previous, no children, etc.). Thankfuilly, the Calculator grouping can be queried to see if the first child is results text, and if so, notification text can be discarded. This allows unit conversion notification to be announced by checking for another UIA automation ID that is not results text.