Edge: announce address auto-complete.
Reported by a user: in Edge, it is possible to announce auto-comlete for addresses. This wasn't possible until now because that search field was flagged as not supporting automatic selection detection. Thus remove that flag (via overlay class chooser), which then allows text of the addresses to be spoken and brailled (if possible).