Action Center/quick actions: do not allow repetetive announcements for item status changes.
Possibly broken UIA implementation (again): when Action Center is opened more than once, repeated item status property change events are fired. At least this will be logged as a debug message to demonstrate the problem, but when it comes to announcing this, do not repeat it through caching immediate past item status message.