Emoji panel: catch possible attribute error when emoji panel is being closed.
In build 18267, possibly because of IA2 to UIA bridge or something, attribute error is raised when emoji panel is closing. This also causes noticeable freeze for a few seconds (workaround is to reopen emoji panel and go to the desktop right afterwards). Thus catch attribute error exception and return from name change event.