Start menu pixel size: move this live region mitigation to global plugin.
In build 18282, Start menu size text is now part of a new app named StartMenuExperienceHost. This means this live region will be announced at various times, especially when Start menu size changes (in terms of overlal screen real estate). Therefore move the live region mitigation to the global plugin so it can be handled no matter which app the region resides in.