Localization: add skip translation module so messages can be pulled from NVDA Core.
Credit: Alberto Buffalino for skip translation module: with this module, messages that are already present in NVDA Core will be used instead of translators providing add-on specific version (which are often duplicates of Core messages). Not only this lessens burden on translating this add-on, it allows message reuse, as all the translator needs to do is translate Core messages.