WinTenObjs: detect Windows 10 in overlay class chooser and return immediately if an earlier Windows release is in use.
Reported by a user and subsequently confirmed: if one uses this add-on on a Windows 7 or 8.1 computer, NVDA keeps playing error tones while using Windows Explorer. Turns out a COM error from overlay class chooser is to blame, so detect it early and return immediately if the Windows release isn't Windows 10 at all (at the top of the function body).