WinTenObjs: do not initialize nor run overlay class chooser if an unsupported Windows 10 build is in use. Fixes #52.
Consider the following: a user runs a portable copy of nVDA with most recent WinTenApps add-on installed, and the add-on starts up while running on an unsupported version of Windows 10. If this happens, NVDA will not do anything and/or error tones will be heard due to numerous COM errors thrown in overlay class chooser.
Thus detect this problem and and just move on. This is done by defining a minimum release tuple (major, minor, supported build) that the add-on can keep an eye on. Eventually, this will be queried from the manifest if supported in future nVDA releases.