Settings: revise feature update text locator strategy due to introduction of download and instlal link.
It isn't just update history that'll be subject to feature update text locator. Turns out this is applicable for 'download and install now' link introduced in later revisions of build 17134 and later. Thankfully, the UI tree is feature update text -> description -> link. Thus 'label' the link by including the feature update title (not the description).
As part of this change, feature update locator logic has been edited. Instead of checking for update history link, a dedicated conditional will be used to check for links as well. This allows link label issues to be identified and corrected in future releases.