Event tracing: open up new logging possibilities via a new function. Re #53.
Instead of logging all events from all apps while debug logging is on, allow a possibility to trace only an app and/or an event even with lgo level set to info. This will involve new command-line switches specific to this add-on. To allow this possibility, UIA object check and debug flag from log function has been moved to its own method.