WinTenObjs: add ability to track specific events and/or events in a specific app. Re #53.
Add two publicly accessible sets for controling which events will be tracked: an events set and an apps set. This is useful for tracking only some events, or wishing to trouble events coming from a specific process. Also, with this change, if custom events and/or apps set is present, event tracking will be enabled in log levels other than debug.