Search field: remove detection of no controller for search fields.
No controller for search field relied on two things: field value, and the fact that a suggestions list was next to it. When it was initially defined, it was used in People app. Since then, app interface and uIA implementation in People app has changed, resulting in suggestion item raising item selected event, negating the need for this workaround. Thus remove detection of these fields from overlay class chooser; the class is kept for historical purposes.
In consideration ofr people using Windows 10 Version 1803, this will be backported to 19.07.