Detect Windows 10 Version 19H2 July builds (18362.xxxxx).
Windows 10 Version 19H2 is considered a patch update to 1903, thus initially sent to slow ring Insiders as a cumulative update to build 18362 (currently 18362.xxxxx). Due to the same build number in use for now, detect 19H2 through UBR (update build revision), which can be retrieved via Windows Registry, and if 19H2 is indeed in use, write an info log entry.