Windows Search: add searchapp app module (renamed from searchui in 20H1).
In build 18965, SearchUI has been renamed to SearchApp to enforce separation from Cortana (Cortana is its own app now). Because it is similar in appearance and function to old search user interface, create an alias app module for old SearchUI. Because of this, any changes introduced in SearchApp will be kept initially in SearchUI until a decision is made to migrate newer features to the dedicated app module.