SearchUI/SearchApp: remove context menu workaround.
For a number of years context menu in Start menu was treated as a UIA list, with NvDA announcing each item as users arrowed through them. Due to redesign in 2019, this assumption is no longer applicable - items are reported correctly with an additional state information that isn't relevant yet. Thus remove context menu item support for newer Windows 10 feature updates.