Cortana: do not announce user's message by checking for types.
In addition to IndexError, TypeError is raised when trying to announce Cortana's message. Luckily this exception is raised when trying to 'announce' a list item object, which is the case for user messages. Thus catch this and do not announce them, which also resolves an issue where user messages were announced.