Open With: announce open with dialog in Version 2004 and later.
NVDA Core issue 11335: in Version 2004, open With dialog is not announced. This is caused by an IAccesible pane gaining focus, and the dialog has no chance of stealing focus. Further testing shows that this so-called IAccessible object is really a UIA element is disguise. Therefore classify Sheell-Flyout object as a good UIA element but only in Version 2004 and later.