Store: do not announce content download progress if not focused on Store itself.
Now that value cange event will be used to anounce content download progress (at lest in newer Store releases), it might be possible that messages will be announced while using apps other than Store. Therefore do not do this through a combination of Python's 'any' functio and focus ancestors (proceed if and only if at least one of the entries in focus ancestors is a Store element).