Event tracking: catch COM error whyen trying to log Automation Id.
Due to timing issues, COM error is thrown when an element does not present Automation Id. Therefore wrap this inside a try block when trying to log this information as part of event tracking. In case Automation Id could not be fetched, say it is 'not found' for ease of diagnoses.