Modern keyboard/hardware suggestions: manuall add candidate window item to local event handler group when it shows up.
To let hardware input suggestions (now a candidate item overlay class) take care of name change event (for announcing keyboard input suggestions), manually add this class to local UIA event handler group if selective UIA event registration is active. For the purposes of the add-on, it may result in extra verbosity for CJK IME window but the new candidate item behavior mitigates this.