WinTenObjs (Flake8 E501): split long lines.
Split long lines found in WinTenObjs global plugin, including long debug statementss, conditionals, and in case of UIA notification event, function argument definition (arguments are indented twice to differentiate with function body). There is a long line related to search suggestion count but that iwll be dealt with in a future commit as it points to a potential dead code.