Revert EOL 1909 due to NVDA 2020.3 infrastructure.
Originally the add-on declared end of support for November 2019 Update (Version 1909/build 18363) by December 2020. However, after evaluating compatibility with NVDA 2020.3, EOL will be pushed to January 2021. If the original schedule was followed, Version 1909 users would be left with encountering compatibility issues with future NVDA releases in terms of UIA object overlay class chooser (if running on 2020.2 or earlier, a patched version of that function will be used, and if the function is changed in future nVDA releases, it will create problems). As a result, upcoming version 21.01 will follow the compatibility plan (NVDA 2020.3 or later) while supporting Windows 10 Version 1909 to reduce issues later on.