WinTenObjs: add type annotations for variables and functions.
Add type annotations to variables and functions that are defined in the add-on. Not all functions are annotated (event_* family are not annotated as they come from NVDA Core). Note that uppercase versions of built-in types are imported to satisfy Mypy for now - Python 3.9 includes annotations compiler flag to let people use lowercase type names.