Settings: only include download link label if the link is next to the update label text.
Consider the following: there is an optional quality update and one or more driver updates. If so, a 'View all updates' button will be present next to quality update download link. This results in quality update label being included in the 'view updates' link, making the label verbose. Therefore include quality/feature update label to 'download and install' lnk if the link is next to the actual label text.