Calculator: remove bulk of the app module as NVDA 2020.4 includes dedicated modern Calculator app module.
One of the oldest parts of Windows 10 App Essentials is now part of NVDA: modern Calculator. Now that NVDA 2020.4 includes this app module, remove bulk of the app module from the add-on apart from handling history and memory items and graphing calculator in recent releases. Also, as part of this commit, remove typing module import from Calculator as it is no longer needed.