Modern keyboard: ignore bulk of element selected and window open events if running on build 21000 and later.
As input experience window uses a different user interface (at least combined emoji panel and clipboard history does), the bulk of element selected and window open events are not applicable. Therefore return early if running on build 21000 (specifically, build 21296) and later.
A subtle benefit of this change is that the add-on can work better with modern IME support introduced in NVDA 2020.4, in that element selected and window open events define specific instructions to handle modern IME items (possibly a later commit to remove duplicates).