Modern keyboard: define a temporary wrapper for winVersion.getWinVer function introduced in NVDA 2021.1.
NVDA 2021.1 introduces winVersion.getWinVer and Windows 10 build constants. To support old and new NVDA releases, define a wrapper for winVersion.getWinVer and Windows 10 build constatns for older NVDA releases (getWinVer wrapper returns WinVersion object in NVDA 2021.1, whereas it will return build integer in older releases). Note that this wrapper is temporary and will be removed in a future add-on release when the add-on will declare NVDA 2021.1 as minimum version supported.