Modern keyboard: set NvDA's focus object to input experience panel to let arrow keys work to select emojis.
Consider the following: in build 20200 and later, a user is focused on an edit field and presses Windows+Period to open emoji panel. Pressing arrow keys is supposed to announce emojis but NVDA is not announcing it. This was traced to editable text object where arrow keys are used by the focused edit field object in NVDA Core for cursor navigation. Since emoji panel is an overlay, emojis are not set as focusable, therefore arrow keys are still used by edit field object. Therefore set NVDA's focus to input experience panel when it opens so arrow keys can be used to select emojis.