Modern keyboard/emoji panel: redirect focus to the actual system focus when modern emoji panel closes.
When modern meoji panel (part of input experience panel in build 20200 and later) closes, focus moves to the panel window itself instead of system focus. Therefore use a gain focus event to redirect focus to the actual control with system focus. This works best with emoji panel portion (clipboard history doesn't do this, although a workaround with a script is next).