Modern keyboard/window open event: move first child (top window) check to above all other parts of the event handler.
There is no need to perform things such as adding elements to local event handler group if the top-level window is gone (None). Therefore move first child check to the top of the window open event handler - right aftre assigning a local variable to obj.firstChild. In a way, this is subtle optimization in that it saves a selective event registration config check.