Modern keyboard/name change event: remove Version 1803 or later condition when announcing clipboard history items.
Back when 1709 and 1803 were supported in name change event handler, clipboard history items said odd things under some circumstances, and as a result, NVDA was told to return from event handler if running on Version 1809 or later. now that 1709 and 1803 are no longer supported, there is no need for a version check when performing clipboard history item announcement check.