Modern keyboard: remove winversion.getWinVer function definition.
A custom version of getWinVer function was needed when working with Windows 10 version constants. Since the only place where build number is checked is for combined emoji panel in build 21296 and later, and since that uses sys.getwindowsversion tuple, there is no need for winVersion and getWinver function definition for now. winvErsion.getWinVer will return with NVDA 2021.1 era.