Reset This PC (System Settings Admin Flow): add header text to local event handler group when the dialog first opens.
With selective UIA event registration on, header text changes were not announced as Reset This PC was being set up. Thankfully the dialog fires a window open event, which can be used to set up event hadnler group support for the header text. Note that the first child of the Reset This PC dialog is a title bar, and the header text is next to it. However simple first child of the dialog points to header text. Therefore use simple first child unless the situation changes in future Windows 10 releases.