Modern keyboard/element selected event: remove dedicated Automation Id variable.
In old add-on releases, a dedicated Automation Id varialbe was used as part of checking for specific Id's for the object. As of June 2021, the only place the dedicated variable is used is checking for candidates list (old emoji panel from 1709 and 1803 days), therefore replace the dedicated variable with obj.UIAAutomationId.