Microsoft Store: separate download title and progress messages, use UIA's notion of first child object when fetching progress message.
Separate download title and progress message into dedicated variables for readability. Also, as an optimization bonus, use UIA's notion of first child element when fetching download progress message instead of using simple first child (this results in improved performance - up to three times faster). Note that due to use of UIA's notion of child elements, this will need to be edited if the Store's download UI changes in the future.