Microsoft Store/value change event: remove try/except block.
In the old days, in the midst of announcing odwnload progress, AttributeError was logged when one of the children of the instlal control element became missing. Recent Store releases appears to have resolved it somewhat (or might be due to UIA changes or comtypes). Therefore remove the try/except block.