Microsoft Store/value change event: announce download progress if install control UI is shown.
Optimization: value change event is fired regardless of UI being shown, and turns out install control does raise this even if the UI is not shown. Therefore limit download progress announcement to presence of the download UI at the cost of missing some announcements (notably, 'working' is announced when the UI is gone).