Mail: handle controller for event for Mail word document root window.
As Root Focus Control element was found in Mail app only, there is no need for WinTenObjs global plugin to define search field for the root Word document control in order to handle controller for event. Therefore transfer the Mail specific controller for event from WinTenObjs to Mail app module (specifically, not relying on focused state of the element that raised controller for event, in this case, Mail document root control).