Modern keyboard: candidate item.last child is the candidate text in Windows 11.
In Windows 11, it is not the first child of the candidate item that raises name change event - rather, there are two children, and the last child is the actual text (for some odd reason). Therefore modify the traversal route used when adding elements to evnet handler group, which will then let NvDA announce hardware keyboard input suggestions.