Settings: gather XAML class name label control descriptors under one list.
Some controls in Settings app are labeled with XAML class names 9SystemSettings_*), and so far, previous.name is the actual label. These include development mode label in Windows 10 2004/20H2/21H1, and Microsoft Account controls in Windows 11. Because the way to obtain correct labels are the same, gather XAML class names under one roof (list) and check control names against this list, and if the names match, use the name of the previous object as its label.