Suggestions list view: transfer bulk of layout invalidated event handler to a new suggestions list view class.
Use a dedicated suggestions list view class to deal with layout invalidated event. As a first step, transfer bulk of layout invalidated event found in the global plugin to the new SuggestionsListView class, substitutig 'obj' for 'self'. The event still relies on focused search field for actual suggestions count announcement, wihch is done for backward compatibility for a time (in the end, the list view itself will be responsible for suggestions count output).