Modern keyboard: ignore modern keyboard active flag on Windows 11.
In Windows 11, to browse all dmojis, one would select 'see more emojis' button. This opens a grid of emojis, and one would use Tab to move between categories, akin to classic emoji panel (Windows 10 1709 and 1803). However, this is preceeded by a state change event coming from the first selected emoji item which is off-screen. This causes modern keyboard active flag to be cleared, causing NVDA to think that element selected event should not proceed when in fact it should. Therefore only check modern keyboard active flag if this is a Windows 10 release (investigate relevance of modern keyboard active flag in a future release).