Modern keyboard: remove combined meoji panel/clipboard history handler.
In newer Windows 11 21H2 (build 22000) revisions, focus moves to emoji panel search field when the panel opens. Therefore it is unnecessary to provide window open event handler for the combined emoji panel/clipboard history. Although this means seelcted emoji will not be announced in some cases, it makes the experience consistent with Narrator.