Modern keyboard/Windows 10 emoji panel flags: remove recently selected flag from window opened event.
Recently selected flag is used to avoid repeating emojis when the panel opens. This used to be the case in early emoji panel releases but not so much since Windows 10 2004 after emoji panel was redesigned (NVDA now queues element selected event when emoji panel opens as opposed to responding to name change event). Therefore remove this flag from the app module.