Modern keyboard/Windows 10 emoji panel flags: remove recently selected flag check from name and state change events.
Recently selected flag is cleared when emoji panel closes. This was a bit difficult to do in recent emoji panel releases as name change is not fired when emoi panel closes without moving to other items. As a workaround, state and name changes were used to determine whether or not recently selected flag should be cleared. Since the only place this flag is used is name change event defined in NVDA Core (announcing anything name change event says), there is no need to handle this flag from the add-on version. Therefore remove this flag check from both name and state change events, making these events simpler in the process.