App modules/control types refactor: revert to pre-refactor code path (controlTypes.ROLE_*/STATE_*).
Depending on NVDA version, controlTypes.ROLE_*/STATE_* will refer to integers (2021.1) or enum (2021.2). Since the old style (integers) will be kept until NVDA 2022.1, and a future version of this add-on will require the new style (2021.2) anyway, it makes sense to use the old style for now (makes it easier to use the refactored paths through search and replace). As part of code revision, remove try/except block and add a note about removal of role/state integers in NVDA 2022.1.