Calculator (Windows 11): add Windows 11 version of Calculator app.
Calculator was redesigned in Windows 11 with new visuals and rewritten in C#. Initial tests indicate that notification event needs to be edited as the display string is announced all the time (again).
Note: this is a partial backport of nickel1 with the new app module coming from older version of Calculator app module compatible with older NVDA releases.