Microsoft Store (Windows 10): remove download list item label workaround.
One of the oldest parts of the add-on: when reviewing updates in Microsoft Store, item labels include text such as 'pending' and such, most noticeable in Windows 10 Store. It turns out not only this element has changed in Windows 11 Store, labels do include additional texts to help users understand the status of the focused item. Therefore at the cost of increased verbosity, remove the download list item label workaround.