WinApps/layout invalidated event handler: revert and add a deprecation/duplication notice.
With NVDA adding support for layout invalidated event in the future, there is no need for this event to be logged from this add-on. Therefore deprecate layout invalidated event handler at the cost of code duplication with Event Tracker - layout invalidated event handler will be removed from Windows App Essentials in a future release.