WinApps: do not track and handle UIA layout invalidated event at teh global plugin level.
As NVDA Core includes UIA layout invalidated event support, with the add-on continuing to support it at the suggestions list view overlay class level, there is no need to track and handle this event at the global plugin class level (besides, this event will be logged via Event Tracker add-on). Therefore remove the event handler altogether but not the event Id and the mechanism involved in adding this event to versions of NVDA without this event present.