Settings: add a note on improved name change event in Storage/disk cleanup.
In the past it was necessary to call live region event handler when handling name change event for Storage/disk cleanup screen in Settings app. Thankfully this is no longer needed as recent Windows 10 and NVDA releases are better at handling UIA events. Therefore add a note about this change in preparation for removing name change event handler from Settings app in a future add-on release.