Settings: remove update progress bar value check.
In older Windows 10 releases, update progress bar was shown alongside update progress text, meaning both the announcement and update progress bar output will be heard. In older add-on releases this was countered with a condition that NVDA should not announce update progress if progress value is above 0 so only the update text was announced. Due to redesigned update progress controls in Windows 10 Version 1803 (April 2018 Update/build 17134), this is no longer needed (update progress is next to updates list, with each item housing progress text), and tests indicate that this change carries over to newer Windows 10 releases (not applicable on Windows 11 as update progress controls are different and it is harder to obtain update progress text without resorting to focus changes). Therefore remove the progress bar value condition wen checking to see if live region change should be handled.