WinApps/XAML headings: remove heading element detection and redundant return statement from overlay classes chooser.
XAML heading elements are detected but are not documented in the add-on documentation. Although useful, this feature causes problems in apps such as Calculator, especially since display element is recognized as a heading when it is not really a heading. Therefore remove XAML heading detection, starting with overlay classes chooser. Along with this, remove redundant return statement from suggestions list view detection as it is no longer necessary (return statement was added to return early, which made sense becasue XAML heading detection was right below it, and with the heading detection removed, there is no need for a return statement).