WinApps/UIA notification: announce microphone mute status from anywhere in Windows 11.
One of the new features planned for stable release is microphone mute status command (Windows+Alt+K). First introduced in an Insider Preview build, this is intended to toggle microphone mute in meeting apps such as Zoom and Microsoft Teams. It turns out a UIA notification event is raised to communicate mic status, and the source of this notification in File Explorer. Since notifications are not nanounced from apps other than the focused app (in this case, apps other than File Explorer), inform NVDA to announce mic status from everywhere.