Install tasks: add a note on Version 21H2 releases (Windows 10/Server 2022/11). Re #66.
In preparation for introducing different feature update detection strategy on Windows 10: until Windows 10 Version 21H1 (build 19043), there was one to one mapping between feature update milestones and builds across clients and servers (such as Windows 10 Version 1809/Server 2019/build 17763). This has changed in 21H2 with three releases: Windows 10 (19044), Server 2022 (20348), and Windows 11 (22000). As Microsoft is focusing on Windows 11 development, Windows 10 is in maintenance mode, therefore prepare to use a different strategy to check for supported feature updates: supported builds list if Windows 10 Version 22H2 turns out to be build 19045/vibranium (or a build lower than 20348). In preparation for the new strategy, add a note about 21H2 releases in install tasks so it can serve as a reminder.