Calculator: add a flag preventing UIA notification event from proceeding if no results command is pressed.
With the introduction of UIA notification event, commands such as number row keys would tell Calculator to update the results display and instruct UIA library t use notification event to present information to NVDA. In its beginning, only a few commands were defined where NVDA would announce resluts upon gesture/script execution. However, it was pointed out in 2022 that this prevents NVDA from announcing the calculator display when keys such as equals (=) or plus (+) are pressed, more so when performing scientific operations (see a eries of previous comits as to what is going on).
Therefore, a completely different design is being proposed: no results flag. This flag will be enabled if a command marked as 'no results' is executed, typically number row keys. This flag, to be checked from UIA notification handler, will let NVDA know that result should not be announced i.e. event should not propagate to the default handler (speech and braille output). As a first step, a dedicated flag is hereby defined to control no results announcement behavior, with a script defining no results command to be added in the next commit.