Calculator: deprecate additional result shortcuts list and add explanatory comments.
As of this commit, additional result shortcuts list and bindGesture call in the constructor are deprecated, to be removed in a future commit when 'do not announce results' flag approach is operational i.e. released and proven to work in stable version of the add-on. Also, add explanatory comment on the new flag approach.