WinApps/search result details (backport): backport search result details announcement fix to copper 1 without a new app module.
Due to UI redesign in 2019. Start menu search result details are not announced apart from the top search result. Thankfully results are the grandchildren of the actual results list. Therefore check obj.parent.parent.
This is a backport of result details announcement fix that does not involve a new app module. This is done by using obj.appModule.appModuleName (searchui) from the global pluin's overlay classes chooser. With this in place, it is equivalent to the dedicated app modue method.